# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Photographer Private CRM API
  version: 0.1.0
  description: 多账号自助认证；所有资源端点隐式按当前账号过滤，客户端永不传 account_id（ADR-001）。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/capabilities:
    get:
      tags: [email-account-access]
      operationId: getAuthCapabilities
      summary: 读取公开注册能力
      security: []
      responses:
        "200":
          description: 当前公开注册能力；不得缓存
          headers:
            Cache-Control:
              schema:
                type: string
              description: 固定 no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCapabilities"
        "500":
          $ref: "#/components/responses/Internal"
  /auth/register:
    post:
      tags: [email-account-access]
      operationId: register
      summary: 注册 pending_verification 账号并尝试发送验证邮件
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email:
                  type: string
                  maxLength: 254
                password:
                  type: string
                  minLength: 12
                  maxLength: 72
      responses:
        "202":
          description: 固定公开结果；不得据此判断账号或投递状态
          headers:
            Cache-Control:
              schema:
                type: string
              description: 固定 no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationDispatch"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/RegistrationDisabled"
        "500":
          $ref: "#/components/responses/Internal"
  /auth/email/resend:
    post:
      tags: [email-account-access]
      operationId: resendVerification
      summary: 以 generic outcome 重发验证邮件
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email]
              properties:
                email:
                  type: string
                  maxLength: 254
      responses:
        "202":
          description: 固定公开结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationDispatch"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /auth/email/verify:
    post:
      tags: [email-account-access]
      operationId: verifyEmail
      summary: 消费 verification／legacy claim token 并建立首个 session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: 已验证并建立 session
          headers:
            Set-Cookie:
              schema:
                type: string
              description: __Host-crm_refresh；HttpOnly、Secure、Strict、Path=/、无 Domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "400":
          $ref: "#/components/responses/InvalidOrExpiredToken"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /auth/login:
    post:
      tags: [email-account-access]
      operationId: login
      summary: 邮箱密码登录并建立 refresh session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email:
                  type: string
                  maxLength: 254
                password:
                  type: string
                  minLength: 12
                  maxLength: 72
      responses:
        "200":
          description: 登录成功
          headers:
            Set-Cookie:
              schema:
                type: string
              description: __Host-crm_refresh；HttpOnly、Secure、Strict、Path=/、无 Domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: forbidden（Origin）或 email_verification_required（密码正确但未验证）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /auth/refresh:
    post:
      tags: [email-account-access]
      operationId: refresh
      summary: 使用 HttpOnly refresh cookie 轮换 session generation
      security: []
      responses:
        "200":
          description: refresh 成功
          headers:
            Set-Cookie:
              schema:
                type: string
              description: rotated __Host-crm_refresh cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          description: unauthorized；同时清除 refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Max-Age=0 且 Expires 为过去时间
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/Internal"
  /auth/logout:
    post:
      tags: [email-account-access]
      operationId: logout
      summary: 幂等撤销 refresh family 并清 cookie
      security: []
      responses:
        "204":
          description: 已退出；missing／expired／revoked 同样成功
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Max-Age=0 且 Expires 为过去时间
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/Internal"
...
  /me:
    get:
      tags: [email-account-access]
      operationId: getMe
      summary: 当前账号信息（永不含 password_hash）
      responses:
        "200":
          description: 当前账号
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /customers:
    post:
      tags: [customers, customer-core]
      operationId: createCustomer
      summary: 30 秒建档端点（仅 3-4 项必填）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name, channel, identities]
              properties:
                display_name:
                  type: string
                channel:
                  $ref: "#/components/schemas/CustomerChannel"
                referrer_customer_id:
                  type: string
                  description: channel=referral 时必填；介绍人须 active 且本账号可见，非 active 或跨账号按 404 处理
                identities:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [platform, handle]
                    properties:
                      platform:
                        $ref: "#/components/schemas/SocialPlatform"
                      handle:
                        type: string
                      remark:
                        type: string
      responses:
        "201":
          description: 已建档
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
    get:
      tags: [customers, customer-core]
      operationId: listCustomers
      summary: 客户列表（q 匹配 display_name/real_name/phone/identity.handle）
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: channel
          in: query
          schema:
            $ref: "#/components/schemas/CustomerChannel"
        - name: status
          in: query
          description: 缺省 active；兼容单值/all，也接受逗号分隔的 active/archived/merged 状态集合；重复值、未知值或 all 与其他值混用均为 400；过滤发生在分页前
          schema:
            type: string
            pattern: "^(all|(active|merged|archived)(,(active|merged|archived))*)$"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 分页客户列表
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/CustomerListItem"
                  total:
                    type: integer
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /customers/{id}:
    get:
      tags: [customers, customer-core]
      operationId: getCustomer
      summary: 客户详情（含 identities、notes 倒序、referrer 摘要）
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 客户详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
    patch:
      tags: [customers, customer-profile-complete]
      operationId: updateCustomer
      summary: 渐进补全任意字段；{status:archived} 即归档（§4.2）
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  description: 不可清空（不接受 null）
                real_name:
                  type: string
                  nullable: true
                  description: 显式传 null 清空
                phone:
                  type: string
                  nullable: true
                  description: 显式传 null 清空
                birthday:
                  nullable: true
                  allOf:
                    - $ref: "#/components/schemas/Birthday"
                  description: 显式传 null 清空
                channel:
                  $ref: "#/components/schemas/CustomerChannel"
                referrer_customer_id:
                  type: string
                  description: channel=referral 时必填；介绍人须 active 且本账号可见，非 active 或跨账号按 404 处理；不得为本客户自身（400）
                status:
                  type: string
                  enum: [active, archived]
                  description: merged 是 merge 端点专属终态，不可经 PATCH 设置
      responses:
        "200":
          description: 更新后的客户
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: customer_merged（merged 客户档案只读）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/Internal"
  /customers/{id}/avatar:
    put:
      tags: [customers, customer-avatar]
      operationId: putCustomerAvatar
      summary: 条件设置或替换客户头像；active/archived 可写，merged 恒为 customer_merged
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/AvatarIfMatch"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 设置后的客户；same-content 完整对象重放为 no-op
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: customer_merged | avatar_revision_conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/Internal"
    delete:
      tags: [customers, customer-avatar]
      operationId: deleteCustomerAvatar
      summary: 条件移除客户头像；merged 仅开放本 cleanup-only 操作
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/AvatarIfMatch"
      responses:
        "200":
          description: 移除后的客户；已无头像时为 no-op
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: avatar_revision_conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/Internal"
  /customers/{id}/avatar/content:
    get:
      tags: [customers, customer-avatar]
      operationId: getCustomerAvatarContent
      summary: 经账号鉴权读取当前强版本头像内容
      parameters:
        - $ref: "#/components/parameters/Id"
        - name: v
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/AvatarVersion"
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
            pattern: '^"sha256-[0-9a-f]{64}"$'
      responses:
        "200":
          description: 完整性已验证的当前头像字节
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
            Vary:
              schema:
                type: string
            X-Content-Type-Options:
              schema:
                type: string
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "304":
          description: 完整性验证通过且 If-None-Match 命中
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
            Vary:
              schema:
                type: string
            X-Content-Type-Options:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: avatar_version_stale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/Internal"
  /customers/{id}/identities:
    post:
      tags: [customers, customer-profile-complete]
      operationId: addCustomerIdentity
      summary: 追加社交身份
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, handle]
              properties:
                platform:
                  $ref: "#/components/schemas/SocialPlatform"
                handle:
                  type: string
                remark:
                  type: string
      responses:
        "201":
          description: 已追加
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialIdentity"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: customer_merged（merged 客户档案只读）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/Internal"
  /customers/{id}/identities/{identity_id}:
    delete:
      tags: [customers, customer-profile-complete]
      operationId: deleteCustomerIdentity
      summary: 删除社交身份
      parameters:
        - $ref: "#/components/parameters/Id"
        - name: identity_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 已删除
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: customer_merged（merged 客户档案只读）| last_identity（非 merged 客户至少保留 1 个社交身份）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/Internal"
  /customers/{id}/notes:
    post:
      tags: [customers, customer-profile-complete]
      operationId: addCustomerNote
      summary: 追加客户备注
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 500
      responses:
        "201":
          description: 已追加
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerNote"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: customer_merged（merged 客户档案只读）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/Internal"
  /customers/{id}/merge:
    post:
      tags: [customers, customer-profile-complete]
      operationId: mergeCustomer
      summary: 把 source 客户 merge 进当前客户（语义见 §4.2）
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_customer_id]
              properties:
                source_customer_id:
                  type: string
      responses:
        "200":
          description: merge 后的 target 客户
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: merge_conflict（source 或 target 非 active）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/Internal"
  /packages:
    post:
      tags: [packages]
      operationId: createPackage
      summary: 新建套系
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
    get:
      tags: [packages]
      operationId: listPackages
      summary: 套系列表（?status=active 供下单选择）
      parameters:
        - name: status
          in: query
          description: 缺省 active；可显式查 archived/all（§4.3 套系上架/下架语义）
          schema:
            type: string
            enum: [active, archived, all]
            x-enum-varnames:
              - ListPackagesParamsStatusActive
              - ListPackagesParamsStatusArchived
              - ListPackagesParamsStatusAll
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 分页套系列表
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/PackageListItem"
                  total:
                    type: integer
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
...
  /reminders/{id}/done:
    post:
      tags: [reminders, reminder-engine]
      operationId: markReminderDone
      summary: 标记提醒完成
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 更新后的提醒
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /reminders/{id}/dismiss:
    post:
      tags: [reminders, reminder-engine]
      operationId: dismissReminder
      summary: 忽略提醒
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 更新后的提醒
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /admin/reminders/scan:
    post:
      tags: [reminders, reminder-engine]
      operationId: scanReminders
      summary: 提醒扫描唯一手动触发入口（每日定时与手动共用，鉴权同 §4.1）
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                  description: 缺省=账号时区今日
      responses:
        "200":
          description: 扫描结果计数
          content:
            application/json:
              schema:
                type: object
                required: [created, skipped, auto_dismissed]
                properties:
                  created:
                    type: integer
                  skipped:
                    type: integer
                  auto_dismissed:
                    type: integer
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationFailed:
      description: 400 validation_failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RateLimited:
      description: 429 rate_limited
      headers:
        Retry-After:
          schema:
            type: integer
          description: 固定秒数；不暴露 limiter 维度
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RegistrationDisabled:
      description: 503 registration_disabled
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Internal:
      description: 500 internal
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    AuthCapabilities:
      type: object
      additionalProperties: false
      required: [public_registration_enabled]
      properties:
        public_registration_enabled:
          type: boolean
    VerificationDispatch:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [verification_required]
    AccessTokenResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          enum: [600]
...